Add per page validation
This implementation is a completion of the intent suggested by the
commented (out) code at original lines 30 and 74.  The feature uses
submit to validate the entire form and checks the AJAX response object
for errors on the current page.  The missing required fields flagged on
subsequent  pages are ignored and cleared on these "multi-page submits".
This feature enforces the validation and does not allow navigation away
from the current page until validation is passed.  A soft validation
option (validation w/out enforcement - error messages only) should be
discussed.